Convert project from MySQL to PostgreSQL (compatible with Neon)
- Rewrote all 11 SQL migrations to PostgreSQL syntax (SERIAL, BIGSERIAL,
  SMALLINT, CHECK constraints, JSONB, STRING_AGG for views)
- INSERT IGNORE replaced with INSERT ... WHERE NOT EXISTS
- Engine/charset/collation/backtick syntax removed
- Database.php: driver-agnostic DSN builder, INSERT with RETURNING id
  for PostgreSQL, auto updated_at on UPDATE
- config/database.php: added DB_DRIVER env var, dynamic port default
- Product.php: GROUP_CONCAT → STRING_AGG
- LoginAttempt.php: DATE_SUB(NOW(), INTERVAL) → PHP timestamp calculation
- .env.example added with Neon-compatible template

+++ migrations/010_create_views_dashboard.sql
@@ -28,7 +28,7 @@ SELECT
 
 -- Vista: Productos por categoría (para gráficos)
 CREATE OR REPLACE VIEW v_products_by_category AS
-SELECT c.id, c.name, COUNT(p.id) AS product_count
+SELECT c.id, c.name, COUNT(p.id)::INTEGER AS product_count
 FROM categories c
 LEFT JOIN products p ON p.category_id = c.id AND p.is_active = 1
 WHERE c.is_active = 1
